requirements.txt
albumentations==0.5.2
accelerate==0.30.1
braceexpand==0.1.7
diffusers==0.28.0
easydict==1.9.0
einops==0.7.0
gradio==4.32.2
gradio-imageslider==0.0.20
hydra-core==1.1.0
huggingface-hub==0.23.0
imageio==2.34.1
imageio-ffmpeg==0.5.0
imgaug==0.4.0
joblib==1.4.2
kornia==0.5.0
matplotlib==3.8.2
numpy==1.26.2
omegaconf==2.1.2
opencv_python==4.8.1.78
packaging==23.2
pandas==2.2.2
Pillow==10.3.0
pytorch_lightning==1.2.9
PyYAML==6.0.1
pygltflib==1.16.1
scikit-image==0.25.0
scikit_learn==1.6.0
scipy==1.11.4
six==1.16.0
spaces==0.28.3
tabulate==0.9.0
tensorflow==2.18.0
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.36.1
trimesh==4.0.5
timm==0.9.12
webdataset==0.2.100
wldhx.yadisk-direct==0.0.6
xformers==0.0.24

